104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,22 +7,22 @@
  */
 int main(void)
 {
-	unsigned int fib1 = 1, fib2 = 2, next_fib;
-	int count;
+    unsigned long fib1 = 1, fib2 = 2, next_fib;
+    int count;
 
-	printf("%u, %u", fib1, fib2);
+    printf("%lu, %lu", fib1, fib2);
 
-	for (count = 3; count <= 98; count++)
-	{
-		next_fib = fib1 + fib2;
+    for (count = 3; count <= 98; count++)
+    {
+        next_fib = fib1 + fib2;
 
-		printf(", %u", next_fib);
+        printf(", %lu", next_fib);
 
-		fib1 = fib2;
-		fib2 = next_fib;
-	}
+        fib1 = fib2;
+        fib2 = next_fib;
+    }
 
-	printf("\n");
+    printf("\n");
 
-	return (0);
+    return (0);
 }
